refactor: extract environment config boundary (#73)
